clean-up: separate utils from pipeline
