+++ mozzarellm/analyzer.py
@@ -12,7 +12,7 @@ import pandas as pd
 from tqdm import tqdm
 
 from .models import AnalysisResult, ClusterResult
-from .clients.LLM_api_clients import create_client
+from .clients.llm_api_clients import create_client
 from .utils.llm_analysis_utils import process_cluster_response
 from .utils.prompt_factory import make_cluster_analysis_prompt
 from .utils.retrieval import local_knowledge_context_retriever
