+++ mozzarellm/clients/__init__.py
@@ -1,0 +1,1 @@
+"""Client implementations for external services (LLMs, UniProt, etc.)."""
