+++ mozzarellm/pipeline/__init__.py
@@ -1,0 +1,1 @@
+"""Main orchestration for gene cluster analysis"""
